+++ src/bl4ckscor3/bot/bl4ckb0t/misc/SpellingCorrection.java
@@ -1,5 +1,9 @@
 package bl4ckscor3.bot.bl4ckb0t.misc;
 
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+
 import org.pircbotx.hooks.events.MessageEvent;
 
 import bl4ckscor3.bot.bl4ckb0t.core.Bot;
@@ -8,33 +12,30 @@ import bl4ckscor3.bot.bl4ckb0t.util.Utilities;
 
 public class SpellingCorrection
 {
-	//making room for 50 channels, each storing 50 users including their latest message
-	private static String[][] messages = new String[50][50];
-	//storing the position of the channels in the messages array
-	private static String[] channels = new String[50];
+	//<channel, <user#message>>
+	public static HashMap<String, List<String>> storage = new HashMap<String, List<String>>();
 	//needed to check if the message should be added to the array or not (used in Listener.java)
 	public static boolean corrected = false;
 
-	public static void correctSpelling(MessageEvent<Bot> event, String[] split, boolean correctsDifferentUser, String userToCorrect)
+	private static void correctSpelling(MessageEvent<Bot> event, String[] split, boolean correctsDifferentUser, String userToCorrect)
 	{
 		String toReplace = split[1];
 		String replaceWith = split[2];
-		int channelPosition = getChannelPosition(event.getChannel().getName());
 
-		for(String s : messages[channelPosition])
+		for(String s : storage.get(event.getChannel().getName()))
 		{
 			if(s == null)
 				break;
 
 			if(userToCorrect.equals(s.split("#")[0]))
 			{
-				String previousMessage = getLatestMessage(userToCorrect, channelPosition);
-				String correctedMessage = getLatestMessage(userToCorrect, channelPosition).replace(toReplace, replaceWith);
+				String previousMessage = getLatestMessage(userToCorrect, storage.get(event.getChannel().getName()));
+				String correctedMessage = getLatestMessage(userToCorrect, storage.get(event.getChannel().getName())).replace(toReplace, replaceWith);
 
 				if(previousMessage.equals(correctedMessage))
 					return;
 				
-				updateLatestMessage(channels[channelPosition], correctedMessage, userToCorrect);
+				updateLatestMessage(event.getChannel().getName(), correctedMessage, userToCorrect);
 
 				if(correctsDifferentUser)
 					Utilities.chanMsg(event, userToCorrect + " " + L10N.strings.getString("correction.1") + " " + event.getUser().getNick() + " " + L10N.strings.getString("correction.2") + ": " + correctedMessage);
@@ -52,9 +53,11 @@ public class SpellingCorrection
 	public static void updateLatestMessage(String channel, String msg, String username)
 	{
 		int i = 0;
-		int channelPosition = getChannelPosition(channel);
 
-		for(String s : messages[channelPosition])
+		if(!storage.containsKey(channel))
+			storage.put(channel, new ArrayList<String>());
+		
+		for(String s : storage.get(channel))
 		{ 
 			//if the current array position contains no data to replace, stop the loop and add the data
 			if(s == null)
@@ -63,14 +66,14 @@ public class SpellingCorrection
 			//checking for the correct array position to potentially replace the latest message with
 			if(s.split("#")[0].equals(username))
 			{
-				messages[channelPosition][i] = username + "#" + msg;
+				storage.get(channel).set(i, username + "#" + msg);
 				return;
 			}
 
 			i++;
 		}
 
-		messages[channelPosition][i] = username + "#" + msg;
+		storage.get(channel).add(username + "#" + msg);
 	}
 
 	/**
@@ -78,9 +81,9 @@ public class SpellingCorrection
 	 * @param user - The name of the user to get the latest message from
 	 * @return - The latest message from the given user
 	 */
-	private static String getLatestMessage(String user, int channelPosition)
+	private static String getLatestMessage(String user, List<String> messages)
 	{
-		for(String s : messages[channelPosition])
+		for(String s : messages)
 		{
 			if(s.split("#")[0].equals(user))
 				return s.split("#")[1];
@@ -127,7 +130,7 @@ public class SpellingCorrection
 
 				if(split.length == 3 && split[0].equals("s"))
 				{
-					SpellingCorrection.correctSpelling(event, split, true, colon ? message.split(":")[0] : message.split(",")[0]);
+					correctSpelling(event, split, true, colon ? message.split(":")[0] : message.split(",")[0]);
 					corrected = true;
 				}
 
@@ -148,50 +151,5 @@ public class SpellingCorrection
 			else
 				return;
 		}
-	}
-
-	/**
-	 * Returns the position of the given channel in the channels array
-	 * @param channel - The channel to search for
-	 * @return - Position of the given channel in the channels array
-	 */
-	private static int getChannelPosition(String channel)
-	{
-		boolean channelNotFound = false;
-		int channelPosition = 0;
-		
-		try
-		{
-			for(String s : channels)
-			{
-				if(s.equals(channel))
-					break;
-
-				channelPosition++;
-			}
-		}
-		catch(NullPointerException e)
-		{
-			channelNotFound = true;
-		}
-
-		//writing the channel position into the array if the channel has no been found
-		if(channelNotFound)
-		{
-			channelPosition = 0;
-
-			for(String s : channels)
-			{
-				if(s == null)
-				{
-					channels[channelPosition] = channel;
-					break;
-				}
-
-				channelPosition++;
-			}
-		}
-		
-		return channelPosition;
 	}
 }